Add `exportRoomKeys` to `CryptoBackend`.
Element-web calls `exportRoomKeys` on logout, so we need a stub implementation
to get it EW working with the rust crypto sdk.
